remove an ambiguity
We allowed a "while" to be expressed as:

    while <expr> IS <matcher> :
    or
    while <expr>:

The problem is that <expr IS matcher> is itself an expression, so the
first of those whiles overlaps with the second. That is, if I do

   while foo is String:

Then that could either be:

   while <expr=foo> IS <matcher=String>
   or
   while <expr=foo is String>

As it happens, the semantics are exactly the same, so this ambiguity
is just an overlap, and easy to fix!