abstract out the Symbol
Make it an abstract class that knows how to push its values to the stack
or store them from it. This will make it easier to work with fields;
they'll just be symbols that you can look up like anything else, and
once you have them, you can use them to manipulate the stack without
caring about whether you're working with a field or local variable.