simplify the MatcherCompiler
It won't worry about the call site's semantics; it just works with
GOTOs. That means that in an expression context, it'll be something
like:

    <matcher stuff, going to @noMatched if not true>
    bool true
    goto @done
    label @noMatched
    bool false
    label @done

If there are any outer-scope vars (`@varName`), then this is actually
exactly what we want. If there aren't, we can simplify it as a
post-production step if we really want. For instance, if the expression
was:

    is foo :? bar()

then it'll be:

    svar $foo
    call : :bar # returns bool
    gofi @noMatched
    bool true
    goto @done
    label @noMatched
    bool false
    label @done

The pattern from the gofi to the end is easy to find, and always just
simplies to no-op (or if you want to be really strict, a type check
that the top of the stack is True).