fix qualified names
This approach is simpler for a few reasons. First, the grammar doesn't
allow for the possibility of the qualified name ending in an arg
invocation. That's never what we want, and while antlr4 is smart enough
to figure that out, it's more difficult than it needs to be.

Perhaps as importantly, it makes it easier to analyze the name. If
the middle section's length is:

    0: Just a var/field/static method
    1: If it's in a var context (e.g. an assignment), then the middle
       section refers to an instance var, and the IDENT_NAME after the
       middle refers to the field name.
       If it's in a method invocation context, then the middle section
       refers to an instance var and the IDENT_NAME refers to the method
       being invoked.
    2 and up: Not allowed.