some work on StatementCompiler, others
The compilation deps are getting out of hand. I think I need to
(a) create a struct called CompilationContext to unify them, and
(b) create a common testing framework that can instantiate it

Also, for qualifiedIdentNameMiddle: I don't need argsInvocation in these
middle parts yet, so I'm removing it from the grammar.