part 3/3 of unifying what we pass to the compilers
Unified both of the tests to use the same base class. The unification
is now complete!

Also, removed a failing test in ParseTest's expression.yaml that
had a long change of qualified name, with an arg invocation in the
middle of it. I removed that from the parser in b4d3e5a because I don't
need it (and can't analyze it yet), and just forgot to remove this test.