remove TODO without doing it
The problem of validating a return statement is harder than just a
boolean. We really want to validate that (a) all returns comply, and
(b) if there is return value, then there's a return in every path.

I think that's more involved than just a flag here, and may be
reasonable to defer altogether until I rewrite the compiler in Effes.