change how FUNC scoping works
According to the evm spec (heh), we should do something like
Module:TypeName, or :TypeName for short. But the actual evm impl just
expects TypeName. That actually makes more sense, as we don't want
modules defining functions on behalf of other modules, so I'm keeping
the "broken" evm behavior and changing the compiler to match.