fix matcher binding in effesj; binary exprs in ef
The matcher bindings was a scoping issue. When we created the scratch
vars within the matcher evaluation scope, we never released them. That
scope has to include the subsequent body (or else bound variables won't
be found), but for the scratch vars, we want that scope to go away.
Without that, when you reference the var name within the body (after
the matcher), you'll be referencing the scratch var, not the real var.

Added some tests for that, and while I was at it, I also cleaned up
some stuff in CompilerTestBase/ResourceReader.

Separately from all that, I have simple expressions working now in
Parser.ef. (My implementation ran into the bug described above, which
is how I found it.)