add the rest of the expressions
- method calls (qualified, unqualified, or static)
- variables (ditto)

Note that this approach to parsing does not allow for chained method
calls: foo(1).bar(2). I'll probably want to refine this at some point,
adding a new level of parsing precedence for method calls just above
this one. That would also allow for things like foo(1).someField,
which this approach does not allow.